feat: Enhance AMM functionality with deposit instruction and error handling
- Added `Deposit` instruction for liquidity addition in `deposit.rs`.
- Updated `ErrorCode` enum to include `InvalidVault` for better error management.
- Changed `index` type from `u8` to `u16` in `AmmConfig` for improved capacity.
- Updated `PoolState` structure to include new bump fields for better state management.
- Enhanced `initialize` function to set new bump values for pool state and vaults.

+++ programs/veerbal_cpmm/src/instructions/create_config.rs
@@ -17,7 +17,7 @@ pub struct CreateAmmConfig<'info> {
 
 pub fn create_amm_config(
     ctx: Context<CreateAmmConfig>,
-    index: u8,
+    index: u16,
     trade_fee_rate: u64,
     creator_fee_rate: u64,
     protocol_fee_rate: u64,
